refactor: rename reducers into bundles
