+++ src/README.md
@@ -2,7 +2,7 @@
 
 This is the source of the client.
 
-It is a [React](https://reactjs.org/) and [Hoodie](http://hood.ie/) based client. State gets handled by [redux](https://redux.js.org/), [redux-thunk](https://github.com/reduxjs/redux-thunk) and [reselect](https://www.npmjs.com/package/reselect).
+It is a [React](https://reactjs.org/) and [Hoodie](http://hood.ie/) based client. State gets handled by [redux-toolkit](https://redux-toolkit.js.org/).
 
 In this directory are the general and setup modules located. Most of the App is in the directories.
 
@@ -10,13 +10,12 @@ In this directory are the general and setup modules located. Most of the App is 
 
 - [`__tests__`](./__tests__) contains all dependencies and integration tests.
 - [`actions`](./actions) contains all redux-actions.
+- [`bundles`]('./bundles) hold and handle state changes. They also export selectors to access state.
 - [`components`](./components) contains all react-components.
 - [`containers`](./containers) contains the glue-code between state and components.
 - [`icons`](./icons) all icons and images.
 - [`network`](./network) contains the UDP-network lair.
 - [`reactors`](./reactors) contains functions that react to state changes and dispatch actions.
-- [`reducers`]('./reducers) update the state by reducing the old state and actions to the new state.
-- [`selectors`](./selectors) contains functions to access and derive state.
 - [`store`](./store) sets up the store.
 - [`types`](./types) TypeScript type declarations that get used in more than one place.
