+++ src/bundles/inventory.ts
